chore(refactor): add refactor helper folder
- add query class to query the model

[Finishes #170844806]
