+++ .coveralls.yml
@@ -1,0 +1,2 @@
+service_name: travis-pro
+repo_token: lvm9ujbgTnsv1bhEjolU2NmdPF9bkDbPD
